+++ CHANGELOG.md
@@ -8,6 +8,8 @@ pip install moldf -U
 
 ## New features
 - Can get a list of chain ids through `PDBDataFrame.chain_list` property, even when two chains are identical as to `chaid_id`s.
+- Can read the `HET`, `HETNAM`, `HETSYN`, and `FORMUL` lines into a `_chem_comp` category now for `read_pdb`.
+The `_chem_comp` category mimics the same category in the `mmCIF` format but is slightly different in that the column names are mostly taken from the `FIELD` specification in the [RCSB's doc](https://www.wwpdb.org/documentation/file-format-content/format33/sect4.html).
 
 ## Fixes